feat(enums): consolidate TaskPriority and AgentStatus to canonical status_enums (#7504)
Consolidates 3 TaskPriority enum definitions:
- subagent_task.py: (str, Enum) with LOW/NORMAL/HIGH/URGENT
- task_queue.py: integer-valued (1-4) for Redis scoring
- orchestrator.py: (Enum) with LOW/NORMAL/HIGH/URGENT

And 2 AgentStatus enum definitions:
- base_agent.py: (Enum) for operational health
- models/agent.py: (str, Enum) for registry lifecycle

All now import from autobot_shared.status_enums:
- TaskPriority → Priority alias (handles string values + numeric conversion)
- AgentStatus (operational) → imported directly
- AgentLifecycleStatus (registry) → aliased as AgentStatus in models

Backward compatibility: task_queue.py deserialization handles both int/string
priority values for old Redis data. All callsites verified.
